requirements.txt
pandas
numpy
scikit-learn
scipy
pyyaml
joblib
dvclive
dvc